refactor: resolveCalculation 함수를 utils → services 폴더로 이동
