+++ src/utils/index.ts
@@ -1,4 +1,3 @@
 export * from './calculate';
 export * from './typeChecker';
 export * from './validator';
-export * from './resolveCalculation';
